requirements.txt
Flask~=3.1.0
pytesseract~=0.3.13
pillow~=11.0.0
Werkzeug~=3.1.3
opencv-python~=4.10.0.84
PyPDF2~=3.0.1
pdf2image~=1.17.0
logging~=0.4.9.6
easyocr~=1.7.2
numpy~=1.26.4
protobuf~=5.28.3
openai~=1.54.4